requirements.txt
django==4.1.4
djangorestframework==3.14.0
pillow==9.3.0
